app-python/requirements.txt
python-keycloak==0.20.0
